wip: enable parallelization
BREAKING CHANGE: By default, tests are now run in parallel. The concurrency is automatically computed based on the number of _physical_ CPU cores detected. This might expose flakiness in tests that previously seemed reliable when running under a less-strained system. If you need to stick to running tests one-at-a-time, pass in the `--concurrency=1` flag.

BREAKING CHANGE: All code specific to GSAP has been removed, in an effort to further generalize this tool. 

What this means in practice is that if you are still running GSAP 2, then `nodecg-screenshot-tester` will no longer automatically detect when a Timeline or Tween is returned by your `entranceMethod`, and it will not wait for it to complete before taking the screenshot.

However, `nodecg-screenshot-tester` _does_ still detect when `entranceMethod` returns a Promise, and waits for that. GSAP 3 now has a `.then` method on all animations, so if you upgrade your codebase to GSAP 3 then you will get this functionality back. Else, you may need to wrap your GSAP 2 timelines in promises in your own codebase.