fix: ensure loads are done before taking screenshot
Turns out there has been a big nasty bug lurking in here for years: `nodecg-screenshot-tester` wouldn't wait for all assets on the page to finish loading before taking the screenshot. This is now fixed, and we'll wait for the network to be idle for at least 500ms before we take the screenshot, ensuring that any in-flight requests have completed.