refactor: remove all GSAP-specific code
BREAKING CHANGE: All code specific to GSAP has been removed, in an effort to further generalize this tool.

What this means in practice is that if you are still running GSAP 2, then nodecg-screenshot-tester will no longer automatically detect when a Timeline or Tween is returned by your entranceMethod, and it will not wait for it to complete before taking the screenshot.

However, nodecg-screenshot-tester does still detect when entranceMethod returns a Promise, and waits for that. GSAP 3 now has a .then method on all animations, so if you upgrade your codebase to GSAP 3 then you will get this functionality back. Else, you may need to wrap your GSAP 2 timelines in promises in your own codebase.